start rework to use IBMi5OSKeyStore
