+++ src/main/java/com/github/ibmioss/dcmtools/DcmExportCertCmd.java
@@ -2,6 +2,7 @@ package com.github.ibmioss.dcmtools;
 
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
@@ -113,10 +114,8 @@ public class DcmExportCertCmd {
             printUsageAndExit();
         }
         try {
-            final File dcmStore = CertUtils.exportDcmStore(logger, opts.isYesMode(), opts.getDcmStore(), opts.getDcmPassword(), null);
-            final KeyStoreLoader loader = new KeyStoreLoader(null, Arrays.asList(dcmStore.getAbsolutePath()), TempFileManager.TEMP_KEYSTORE_PWD, opts.getLabel(), false);
-            final KeyStore keyStore = loader.getKeyStore();
-            final Certificate cert = keyStore.getCertificate(opts.getLabel());
+            KeyStore ks = CertUtils.exportDcmStoreToKeystoreObj(logger, opts.isYesMode(), opts.getDcmStore(), opts.getDcmPassword());
+            final Certificate cert = ks.getCertificate(opts.getLabel());
 
             if (ExportCertOptions.OutputFormat.PEM == opts.getFormat()) {
                 try (BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(file), "UTF-8"))) {
